+++ src/Products/PageTemplates/Expressions.py
@@ -23,21 +23,25 @@ import logging
 from zope.component import queryUtility
 from zope.i18n import translate
 from zope.interface import implements
+from zope.pagetemplate.engine import ZopeEngine as Z3Engine
+from zope.proxy import removeAllProxies
+from zope.tales.expressions import DeferExpr
+from zope.tales.expressions import NotExpr
+from zope.tales.expressions import PathExpr
+from zope.tales.expressions import StringExpr
+from zope.tales.expressions import Undefs
+from zope.tales.pythonexpr import PythonExpr
 from zope.tales.tales import Context
 from zope.tales.tales import ErrorInfo as BaseErrorInfo
 from zope.tales.tales import Iterator
-from zope.tales.expressions import PathExpr, StringExpr, NotExpr
-from zope.tales.expressions import DeferExpr, Undefs
-from zope.tales.pythonexpr import PythonExpr
 from zope.traversing.interfaces import ITraversable
 from zope.traversing.adapters import traversePathElement
-from zope.proxy import removeAllProxies
-from zope.app.pagetemplate.engine import ZopeEngine as Z3Engine
 
 import OFS.interfaces
 from MultiMapping import MultiMapping
 from Acquisition import aq_base
-from zExceptions import NotFound, Unauthorized
+from zExceptions import NotFound
+from zExceptions import Unauthorized
 
 from zope.contentprovider.tales import TALESProviderExpression
 from Products.PageTemplates import ZRPythonExpr